change id from int to long

+++ src/main/java/com/group3/fundmgt/manager/ManagerController.java
@@ -25,7 +25,7 @@ public class ManagerController {
     }
     
     @GetMapping(path="{employeeId}")
-    public Manager getManager(@PathVariable("employeeId") Integer employeeId) {
+    public Manager getManager(@PathVariable("employeeId") Long employeeId) {
         return managerService.getManager(employeeId);
     }
 }
