+++ src/main/java/com/group3/fundmgt/manager/ManagerNotFoundException.java
@@ -5,9 +5,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class ManagerNotFoundException extends IllegalArgumentException {
-    private final Integer employeeId;
+    private final Long employeeId;
 
-    public ManagerNotFoundException(Integer employeeId) {
+    public ManagerNotFoundException(Long employeeId) {
         super("Fund manager with employee id " + employeeId + " not found.");
         this.employeeId = employeeId;
     }
